Remove reference to object_slot.h
This header has been removed in newer versions of sigc++ and the removal causes a build failure for libwfut. Interestingly, the header file was essentially empty (at least in sigc++ version 2.4.1). Since it does not appear that this header served any useful function, the easiest solution seems to be to remove it. I was able to compile libwfut successfully without this line. If no one has any insight as to why this is included, I recommend that we remove it.